Refactor CompanyService and CompanyController classes
Separated the create and update operations into distinct methods in the CompanyService class. Introduced company validation and exception handling for not found and validation errors. Updated corresponding methods in the CompanyController class and renamed test methods accordingly in CompanyServiceTest class.

+++ src/main/java/org/regikeskus/events/service/CompanyService.java
@@ -1,6 +1,8 @@
 package org.regikeskus.events.service;
 
 import lombok.RequiredArgsConstructor;
+import org.regikeskus.events.exception.CompanyNotFoundException;
+import org.regikeskus.events.exception.CompanyValidationException;
 import org.regikeskus.events.model.Company;
 import org.regikeskus.events.repository.CompanyRepository;
 import org.springframework.stereotype.Service;
@@ -33,15 +35,36 @@ public class CompanyService {
     }
 
     @Transactional
-    public Company createOrUpdateCompany(Company company) {
-        if (company.getLegalName() == null || company.getRegistrationCode() == null) {
-            throw new IllegalArgumentException("Company must have a registration code and a legal name");
-        }
+    public Company createCompany(Company company) {
+        validateCompany(company);
         return companyRepository.save(company);
     }
 
     @Transactional
+    public Company updateCompany(Long participantId, Company updatedCompany) {
+        Company company = companyRepository.findById(participantId).orElseThrow(() -> new CompanyNotFoundException(COMPANY_NOT_FOUND_MESSAGE + participantId));
+
+        company.setLegalName(updatedCompany.getLegalName());
+        company.setRegistrationCode(updatedCompany.getRegistrationCode());
+        company.setNumberOfParticipants(updatedCompany.getNumberOfParticipants());
+        company.setPaymentMethod(updatedCompany.getPaymentMethod());
+        company.setAdditionalInfo(updatedCompany.getAdditionalInfo());
+
+        validateCompany(company);
+        return companyRepository.save(company);
+    }
+
+    private void validateCompany(Company company) {
+        if (company.getLegalName() == null || company.getRegistrationCode() == null) {
+            throw new CompanyValidationException("Company must have a registration code and a legal name");
+        }
+    }
+
+    @Transactional
     public void deleteCompany(Long id) {
+        if (!companyRepository.existsById(id)) {
+            throw new CompanyNotFoundException(COMPANY_NOT_FOUND_MESSAGE + id);
+        }
         companyRepository.deleteById(id);
     }
 }
